+++ R/module-template.R
@@ -63,17 +63,20 @@
 #'
 #' @param ...   Additional arguments. Currently, only the following are supported:\cr\cr
 #'
+#'              \code{children}. Required when \code{type = "parent"}. A character vector
+#'              specifying the names of child modules.
+#'
 #'              \code{open}. Logical. Should the new module file be opened after creation?
 #'              Default \code{TRUE}.\cr\cr
-#'
-#'              \code{unitTests}. Logical. Should the new module include unit test files?
-#'              Default \code{TRUE}. Unit testing relies on the \code{testthat} package.\cr\cr
 #'
 #'              \code{type}. Character string specifying one of \code{"child"} (default),
 #'              or \code{"parent"}.\cr\cr
 #'
-#'              \code{children}. Required when \code{type = "parent"}. A character vector
-#'              specifying the names of child modules.
+#'              \code{unitTests}. Logical. Should the new module include unit test files?
+#'              Default \code{TRUE}. Unit testing relies on the \pkg{testthat} package.\cr\cr
+#'
+#'              \code{useGitHub}. Logical. Is module development happening on GitHub?
+#'              Default \code{TRUE}. Setting up GitHub projects relies on the \pkg{usethis} package.\cr\cr
 #'
 #' @return Nothing is returned. The new module file is created at
 #' \file{path/name.R}, as well as ancillary files for documentation, citation,
@@ -113,25 +116,28 @@ setMethod(
   definition = function(name, path, ...) {
     args <- list(...)
 
-    stopifnot((names(args) %in% c('open', 'unitTests', 'type', "children")))
+    stopifnot((names(args) %in% c("children", "open", "type", "unitTests", "useGitHub")))
 
-    open <- args$open
-    unitTests <- args$unitTests
-    type <- args$type
     children <- args$children
+    open <- args$open
+    type <- args$type
+    unitTests <- args$unitTests
+    useGitHub <- args$useGitHub
 
     # define defaults for ... args
-    if (is.null(open)) open <- interactive()
-    if (is.null(unitTests)) unitTests <- TRUE
-    if (is.null(type)) type <- "child"
     if (is.null(children)) children <- NA_character_
+    if (is.null(open)) open <- interactive()
+    if (is.null(type)) type <- "child"
+    if (is.null(unitTests)) unitTests <- TRUE
+    if (is.null(useGitHub)) useGitHub <- TRUE
 
     stopifnot(
-      is(open, "logical"),
-      is(unitTests, "logical"),
-      is(type, "character"),
       is(children, "character"),
-      type %in% c("child", "parent")
+      is(open, "logical"),
+      is(type, "character"),
+      type %in% c("child", "parent"),
+      is(unitTests, "logical"),
+      is(useGitHub, "logical")
     )
 
     path <- checkPath(path, create = TRUE)
@@ -139,14 +145,20 @@ setMethod(
     dataPath <- file.path(nestedPath, "data") %>% checkPath(create = TRUE)
     RPath <- file.path(nestedPath, "R") %>% checkPath(create = TRUE)
 
-    # empty data checksum file
+    ## empty data checksum file
     cat("", file = file.path(dataPath, "CHECKSUMS.txt"))
 
-    # module code file
+    if (isTRUE(useGitHub)) {
+      ## basic .gitignore file for module data
+      gitignoreTemplate <- readLines(file.path(.pkgEnv[["templatePath"]], "data-gitignore.template"))
+      writeLines(whisker.render(gitignoreTemplate), file.path(dataPath, ".gitignore"))
+    }
+
+    ## module code file
     newModuleCode(name = name, path = path, open = open, type = type, children = children)
 
     if (type == "child" && unitTests) {
-      newModuleTests(name = name, path = path, open = open)
+      newModuleTests(name = name, path = path, open = open, useGitHub = useGitHub)
     }
 
     ### Make R Markdown file for module documentation
@@ -288,7 +300,7 @@ setMethod(
     filenameREADME <- file.path(nestedPath, "README.txt")
 
     moduleRmd <- list(
-      author = Sys.getenv('USER'),
+      author = Sys.getenv("USER"),
       date = format(Sys.Date(), "%d %B %Y"),
       name = name,
       path = path
@@ -361,13 +373,17 @@ setMethod("newModuleDocumentation",
 #' @param open  Logical. Should the new module file be opened after creation?
 #'              Default \code{TRUE} in an interactive session.
 #'
+#' @param useGitHub Logical indicating whether GitHub will be used.
+#'                  If \code{TRUE} (default), creates suitable configuration files (e.g.,
+#'                  \file{.gitignore}) and configures basic GitHub actions for module code checking.
+#'
 #' @author Eliot McIntire and Alex Chubaty
 #' @importFrom reproducible checkPath
 #' @export
 #' @family module creation helpers
 #' @rdname newModuleTests
 #'
-setGeneric("newModuleTests", function(name, path, open) {
+setGeneric("newModuleTests", function(name, path, open, useGitHub) {
   standardGeneric("newModuleTests")
 })
 
@@ -375,112 +391,38 @@ setGeneric("newModuleTests", function(name, path, open) {
 #' @rdname newModuleTests
 setMethod(
   "newModuleTests",
-  signature = c(name = "character", path = "character", open = "logical"),
-  definition = function(name, path, open) {
+  signature = c(name = "character", path = "character", open = "logical", useGitHub = "logical"),
+  definition = function(name, path, open, useGitHub) {
     if (!requireNamespace("testthat", quietly = TRUE)) {
       warning('The `testthat` package is required to run unit tests on modules.')
     }
     path <- checkPath(path, create = TRUE)
-    nestedPath <- file.path(path, name) %>% checkPath(create = TRUE)
-    testDir <- file.path(nestedPath, "tests") %>% checkPath(create = TRUE)
-    testthatDir <- file.path(testDir, "testthat") %>% checkPath(create = TRUE)
+    testthatDir <- file.path(path, name, "tests", "testthat") %>% checkPath(create = TRUE)
+    testDir <- dirname(testthatDir)
 
-    # create two R files in unit tests folder:
+    ## create basic local testing structure based on testthat
+    ## -- see ?testthat
+    unitTests <- list(name = name, path = path)
     unitTestsR <- file.path(testDir, "unitTests.R") # source this to run all tests
-    testTemplate <- file.path(testthatDir, "test-template.R")
+    testTemplate <- file.path(testthatDir, "test-template.R") # template for user-defined tests
 
-    # TODO: move this template to inst/templates and use whisker
-    cat("
-# Please build your own test file from test-Template.R, and place it in tests folder
-# please specify the package you need to run the sim function in the test files.
+    moduleUnitTestTemplate <- readLines(file.path(.pkgEnv[["templatePath"]], "unitTests.R.template"))
+    writeLines(whisker.render(moduleUnitTestTemplate, unitTests), unitTestsR)
 
-# to test all the test files in the tests folder:
-test_dir(\"", testthatDir, "\")
+    moduleUnitTestTemplate <- readLines(file.path(.pkgEnv[["templatePath"]], "test-template.R.template"))
+    writeLines(whisker.render(moduleUnitTestTemplate, unitTests), testTemplate)
 
-# Alternative, you can use test_file to test individual test file, e.g.:
-test_file(\"", file.path(testthatDir, "test-template.R"), "\")\n",
-        file = unitTestsR, fill = FALSE, sep = "")
+    ## create basic testing infrastructure using GitHub Actions
+    ## -- see ?usethis::use_github_action and https://github.com/r-lib/actions/tree/master/examples
+    if (isTRUE(useGitHub)) {
+      ghActionPath <- checkPath(file.path(path, name, ".github", "workflows"), create = TRUE)
 
-    ## test template file
-    cat("
-# Please do three things to ensure this template is correctly modified:
-# 1. Rename this file based on the content you are testing using
-#    `test-functionName.R` format so that your can directly call `moduleCoverage`
-#    to calculate module coverage information.
-#    `functionName` is a function's name in your module (e.g., `", name, "Event1`).
-# 2. Copy this file to the tests folder (i.e., `", testthatDir, "`).\n
-# 3. Modify the test description based on the content you are testing:
-test_that(\"test Event1 and Event2.\", {
-  module <- list(\"", name, "\")
-  path <- list(modulePath = \"", path, "\",
-               outputPath = file.path(tempdir(), \"outputs\"))
-  parameters <- list(
-    #.progress = list(type = \"graphical\", interval = 1),
-    .globals = list(verbose = FALSE),
-    ", name ," = list(.saveInitialTime = NA)
-  )
-  times <- list(start = 0, end = 1)
-
-  # If your test function contains `time(sim)`, you can test the function at a
-  # particular simulation time by defining the start time above.
-  object1 <- \"object1\" # please specify
-  object2 <- \"object2\" # please specify
-  objects <- list(\"object1\" = object1, \"object2\" = object2)
-
-  mySim <- simInit(times = times,
-                   params = parameters,
-                   modules = module,
-                   objects = objects,
-                   paths = path)
-
-  # You may need to set the random seed if your module or its functions use the
-  # random number generator.
-  set.seed(1234)
-
-  # You have two strategies to test your module:
-  # 1. Test the overall simulation results for the given objects, using the
-  #    sample code below:
-
-  output <- spades(mySim, debug = FALSE)
-
-  # is output a simList?
-  expect_is(output, \"simList\")
-
-  # does output have your module in it
-  expect_true(any(unlist(modules(output)) %in% c(unlist(module))))
-
-  # did it simulate to the end?
-  expect_true(time(output) == 1)
-
-  # 2. Test the functions inside of the module using the sample code below:
-  #    To allow the `moduleCoverage` function to calculate unit test coverage
-  #    level, it needs access to all functions directly.
-  #    Use this approach when using any function within the simList object
-  #    (i.e., one version as a direct call, and one with `simList` object prepended).
-
-  if (exists(\"", name, "Event1\", envir = .GlobalEnv)) {
-    simOutput <- ", name, "Event1(mySim)
-  } else {
-    simOutput <- myEvent1(mySim)
-  }
-
-  expectedOutputEvent1Test1 <- \" this is test for event 1. \" # please define your expection of your output
-  expect_is(class(simOutput$event1Test1), \"character\")
-  expect_equal(simOutput$event1Test1, expectedOutputEvent1Test1) # or other expect function in testthat package.
-  expect_equal(simOutput$event1Test2, as.numeric(999)) # or other expect function in testthat package.
-
-  if (exists(\"", name, "Event2\", envir = .GlobalEnv)) {
-    simOutput <- ", name, "Event2(mySim)
-  } else {
-    simOutput <- myEvent2(mySim)
-  }
-
-  expectedOutputEvent2Test1 <- \" this is test for event 2. \" # please define your expection of your output
-  expect_is(class(simOutput$event2Test1), \"character\")
-  expect_equal(simOutput$event2Test1, expectedOutputEvent2Test1) # or other expect function in testthat package.
-  expect_equal(simOutput$event2Test2, as.numeric(777)) # or other expect function in testthat package.
-})",
-      file = testTemplate, fill = FALSE, sep = "")
+      moduleRmdYaml <- list(name = name)
+      renderModuleRmdYamlTemplate <- readLines(file.path(.pkgEnv[["templatePath"]],
+                                                         "render-module-rmd.yaml.template"))
+      writeLines(whisker.render(renderModuleRmdYamlTemplate, moduleRmdYaml),
+                 file.path(ghActionPath, "render-module-rmd.yaml"))
+    }
 })
 
 #' Open all modules nested within a base directory